requirements.txt
gunicorn
flask
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
tox
psycopg2-binary
raven
blinker
redis
rq==0.13.0
boto
flask-restplus
enum34
barcodenumber
stripe
requests
newrelic
flask_cors
factory-boy==2.11.1
pytest-factoryboy==2.0.2
APScheduler
backoff
dateparser
sendgrid
mailchimp3
